Move user-related views to user_views
