+++ myapp/api/urls.py
@@ -3,24 +3,24 @@ from rest_framework.routers import DefaultRouter
 from . import views
 from django.conf import settings
 from django.conf.urls.static import static
-from myapp.api.views import views, event_views
+from myapp.api.views import event_views, user_views
 
 router = DefaultRouter()
 router.register(r'events', event_views.EventViewSet, basename="event")
 router.register(r'eventactionrecords', event_views.EventActionRecordViewSet, basename='eventrecord')
-router.register(r'users', views.UserViewSet, basename="users")
-router.register(r'officers', views.OfficerViewSet, basename="officers")
+router.register(r'users', user_views.UserViewSet, basename="users")
+router.register(r'officers', user_views.OfficerViewSet, basename="officers")
 
-router.register(r'inductees', views.InducteeViewSet, basename="inductees")
-router.register(r'outreach', views.OutreachViewSet, basename="outreach")
+router.register(r'inductees', user_views.InducteeViewSet, basename="inductees")
+router.register(r'outreach', user_views.OutreachViewSet, basename="outreach")
 router.register(r'eventtypes', event_views.EventTypeViewSet, basename="eventtype")
-router.register(r'groups', views.GroupsViewSet, basename="groups")
+router.register(r'groups', user_views.GroupsViewSet, basename="groups")
 
 urlpatterns = [
    path("actions/", event_views.EventActionView),
    path("eventactionrecords/pair/<int:event_pk>/<str:other_user_id>/", event_views.EventActionRecordsForEventUserPair),
-   path('profile/', views.UserProfileView.as_view(), name="user_profile"),
-   path('permissions/', views.PermissionsView)
+   path('profile/', user_views.UserProfileView.as_view(), name="user_profile"),
+   path('permissions/', user_views.PermissionsView)
 ]
 
 urlpatterns += router.urls
